refactored hooks into Provider - CalendarData + CalendarUI
