+++ src/providers/CalendarDataProvider.jsx
@@ -1,55 +1,87 @@
 "use client";
 
-import { useEffect, useMemo, useState } from "react";
-import { fetchEvents, fetchAvailabilities, fetchSubjectsWithTutors, fetchTutors } from "../firestore/firebaseFetch";
-import { splitAvailabilities } from "./calendar/availabilityUtils";
+import React, { createContext, useContext, useState, useEffect, useMemo } from 'react';
+import { fetchEvents, fetchAvailabilities, fetchSubjectsWithTutors, fetchTutors, fetchStudentRequests } from '@/firestore/firebaseFetch';
+import { splitAvailabilities } from '@components/calendar/availabilityUtils';
+import { useEmailQueueMonitor } from '@/hooks/useEmailQueueMonitor';
+
+const CalendarDataContext = createContext(null);
+
+export const useCalendarData = () => {
+	const context = useContext(CalendarDataContext);
+	if (!context) {
+		throw new Error('useCalendarData must be used within CalendarDataProvider');
+	}
+	return context;
+};
 
 /**
-* CalendarDataProvider
-* ---------------------
-* Owns ALL data state (events, availabilities, subjects, tutors, students)
-* and exposes them via a render-prop child function. This moves the data
-* state out of the large parent wrapper so the child decides and manages it.
-*
-* Usage:
-* <CalendarDataProvider userRole={userRole} userEmail={userEmail}>
-* {({ allEvents, setAllEvents, availabilities, setAvailabilities, subjects, tutors, students, splitAvailabilitiesData }) => (
-* // ...your calendar UI that consumes these values
-* )}
-* </CalendarDataProvider>
-*/
-const CalendarDataProvider = ({ userRole, userEmail, children }) => {
-    // DATA STATE (moved down from parent)
-    const [allEvents, setAllEvents] = useState([]);
-    const [availabilities, setAvailabilities] = useState([]);
-    const [subjects, setSubjects] = useState([]);
-    const [tutors, setTutors] = useState([]);
-    const [students, setStudents] = useState([]);
+ * CalendarDataProvider
+ *
+ * Manages calendar data fetching and state.
+ * This context rarely changes - only when events/availabilities are added/updated/deleted.
+ * Keeps data separate from UI state to minimize re-renders.
+ */
+export const CalendarDataProvider = ({ children, userRole, userEmail }) => {
+	const [allEvents, setAllEvents] = useState([]);
+	const [availabilities, setAvailabilities] = useState([]);
+	const [studentRequests, setStudentRequests] = useState([]);
+	const [subjects, setSubjects] = useState([]);
+	const [tutors, setTutors] = useState([]);
+	const [students, setStudents] = useState([]);
+	const [loading, setLoading] = useState(true);
 
+	// Fetch all calendar data
+	useEffect(() => {
+		if (!userRole || !userEmail) return;
 
-    useEffect(() => {
-        // Fetch data once the role/email are known
-        fetchEvents(userRole, userEmail, setAllEvents, setAllEvents, setStudents);
-        fetchAvailabilities(setAvailabilities);
-        fetchSubjectsWithTutors(setSubjects);
-        fetchTutors(setTutors);
+		setLoading(true);
+		fetchEvents(userRole, userEmail, setAllEvents, setAllEvents, setStudents, setLoading);
+		fetchAvailabilities(setAvailabilities);
+		fetchStudentRequests(setStudentRequests);
+		fetchSubjectsWithTutors(setSubjects);
+		fetchTutors(setTutors);
+	}, [userRole, userEmail]);
 
-    }, [userRole, userEmail]);
+	// Split availabilities based on events (memoized)
+	const splitAvailabilitiesData = useMemo(() =>
+		splitAvailabilities(availabilities, allEvents),
+		[availabilities, allEvents]
+	);
 
+	// Email queue monitoring (side effect for teachers)
+	useEmailQueueMonitor(userRole);
 
-    // Derive splitAvailabilities here so the parent doesn't have to
-    const splitAvailabilitiesData = useMemo(() => splitAvailabilities(availabilities, allEvents), [availabilities, allEvents]);
+	const value = useMemo(() => ({
+		allEvents,
+		setAllEvents,
+		availabilities,
+		setAvailabilities,
+		studentRequests,
+		setStudentRequests,
+		splitAvailabilitiesData,
+		subjects,
+		tutors,
+		students,
+		loading,
+		userRole,
+		userEmail
+	}), [
+		allEvents,
+		availabilities,
+		studentRequests,
+		splitAvailabilitiesData,
+		subjects,
+		tutors,
+		students,
+		loading,
+		userRole,
+		userEmail
+	]);
 
-    return children({
-        allEvents,
-        setAllEvents,
-        availabilities,
-        setAvailabilities,
-        subjects,
-        tutors,
-        students,
-        splitAvailabilitiesData,
-    });
-}
-
-export default CalendarDataProvider;
+	return (
+		<CalendarDataContext.Provider value={value}>
+			{children}
+		</CalendarDataContext.Provider>
+	);
+};
